requirements.txt
Django==3.1.7

psycopg2==2.8.6
django-redis==4.12.1

django-split-settings==1.0.1
python-decouple==3.4

pillow==8.1.2
django-compressor==2.4
sorl-thumbnail==12.7.0
django-pagedown==2.2.0
